Module forest_cli_shared::cli

Structs

CalibnetSnapshotFetchConfig

CliOpts

CLI options

Client

Config

DaemonConfig

Structure that defines daemon configuration when process is detached

LogConfig

LogValue

MainnetSnapshotFetchConfig

SnapshotFetchConfig

Enums

ConfigPath

Statics

FOREST_VERSION_STRING

Functions

chain_path

Gets chain data directory

check_for_unknown_keys

cli_error_and_die

Print an error message and exit the program with an error code Used for handling high level errors such as invalid parameters

db_path

Gets database directory

default_snapshot_dir

is_aria2_installed

Checks whether aria2c is available in PATH

snapshot_fetch

Fetches snapshot from a trusted location and saves it to the given directory. Chain is inferred from configuration.

to_size_string

convert BigInt to size string using byte size units (i.e. KiB, GiB, PiB, etc) Provided number cannot be negative, otherwise the function will panic.