Module forest_cli_shared::cli

Structs

CalibnetSnapshotFetchConfig

CliOpts

CLI options

Client

Config

DaemonConfig

Structure that defines daemon configuration when process is detached

LogConfig

LogValue

MainnetSnapshotFetchConfig

SnapshotFetchConfig

Statics

FOREST_VERSION_STRING

Functions

cli_error_and_die

Print an error message and exit the program with an error code Used for handling high level errors such as invalid parameters